+++ src/eslint.js
@@ -56,7 +56,7 @@ module.exports = {
     const installDependenciesTaskName = `${taskName}-install-dependencies`;
     npmInstall.register(gulp, installDependenciesTaskName, {
       dependencies: this.dependencies
-    });
+    }, globals);
 
     const validate = config => {
       const eslint = require('gulp-eslint');