+++ src/nsp.js
@@ -27,6 +27,8 @@ module.exports = {
    * @param {external:gulp} gulp The gulp instance to use
    * @param {string} taskName The name of the task
    * @param {module:tasks/nsp.Parameters} parameters The parameters of the task
+   * @param {Object} [globals] The global configuration store of the tasks
+   *                           Globals are used to set up defaults
    *
    * @example
    * register(
@@ -37,11 +39,11 @@ module.exports = {
    *   }
    * );
    */
-  register(gulp, taskName, parameters) {
+  register(gulp, taskName, parameters, globals) {
     const installDependenciesTaskName = `${taskName}-install-dependencies`;
     npmInstall.register(gulp, installDependenciesTaskName, {
       dependencies: this.dependencies
-    });
+    }, globals);
 
     gulp.task(taskName, [installDependenciesTaskName], (done) => {
       const nsp = require('nsp');