+++ src/retire.js
@@ -28,6 +28,8 @@ module.exports = {
    * @param {external:gulp} gulp The gulp instance to use
    * @param {string} taskName The name of the task
    * @param {module:tasks/retire.Parameters} parameters The parameters of the task
+   * @param {Object} [globals] The global configuration store of the tasks
+   *                           Globals are used to set up defaults
    *
    * @example
    * register(
@@ -36,11 +38,11 @@ module.exports = {
    *   { executablePath: path.join(__dirname, 'node_modules/.bin/') }
    * );
    */
-  register(gulp, taskName, parameters) {
+  register(gulp, taskName, parameters, globals) {
     const installDependenciesTaskName = `${taskName}-install-dependencies`;
     npmInstall.register(gulp, installDependenciesTaskName, {
       dependencies: this.dependencies
-    });
+    }, globals);
 
     gulp.task(taskName, [installDependenciesTaskName], done => {
       const config = buildConfig(parameters);